requirements.txt
ragas>=0.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
datasets>=2.0.0
allure-pytest==2.15.0